correct priceLevel datatype in PlaceDTOs

+++ backend/mad-maps/src/main/java/com/codecool/madmaps/DTO/Place/PlaceDTO.java
@@ -4,5 +4,5 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
-public record PlaceDTO(String placeId, String name, double rating, double priceLevel, List<String> openingHours) {
+public record PlaceDTO(String placeId, String name, double rating, int priceLevel, List<String> openingHours) {
 }
